chore: typescript-app init test fails during release (#9)
The typescript-app template installs the cdk8s-cli in the project as a dev dependency. In non-release builds, the module version is `0.0.0` and sadly we actually published a `0.0.0` once, so this is the version being installed.

In release builds, the version is bumped to a new, non-released version, and then `cdk8s init` fails.

To address this, an environment flag was added `CDK8S_TEST` which is set to true during this test which causes the `cdk8s init` to use the executing cli instead of installing it through npm.

Also: upgrade projen.