fix(import): cannot import props that have non-standard capitalization
json2jsii now emits `toJson()` functions for all generated
structs which can be used to convert data back to the original
schema.

This means that we can now import APIs and CRDs even if their
property names are not properly camel cased, use underscores
or hyphens, etc.

Improve test harness to also take a snapshot of the generated typescript
output