feat: include CLI information for long import times (#53)
I noticed that imports in non-typescript languages sometimes take a long time (upwards of a minute). There's nothing in the CLI that tells you this operation should be long, and almost nothing is output in terms of progress by the jsii transpiler currently, so adding this message might improve the user experience a little bit.