fix: error displays after manifests are successfully generated (#176)
PR addresses the following issue:

When cdk8s synth generates files in sub-folders of the output folder, the cli thinks that no files where generated and prints 'No manifests synthesized' in the console. This is because the CLI is only looking for files in the output directory itself and not its subdirectories.

The PR changes the CLI to search for files both in the output folder and its sub-folders