feat: introduce v2.x (#278)
Since we have two major versions of `cdk8s`, we need the cli to support initializing templates that use either one. 

We could have done this by adding a `--cdk8s-version` flag to the cli and thats it. However:

- We want the default to be `cdk8s` v2, and if we change the default, thats a breaking change.
- Keeping `cdk8s-cli` at v1 is going to cause some confusion with users since other cdk8s versions will be v2.

Granted, introducing a new release branch creates a maintenance burden, but I feel we can handle it with automatic backports for a while, like we do with [`cdk8s-core`](https://github.com/cdk8s-team/cdk8s-core/pull/482) and `cdk8s-plus`. 
Signed-off-by: iliapolo <epolon@amazon.com>