fix: typescript app template does not compile on Windows
cdk8s-team/cdk8s-cli#362 made the `compile` and `watch` scripts of the
`typescript-app` template use a relative path to the `tsc` compiler,
however such a form does not work as intended on Windows...

Instead of doing this, explicitly declare a `devDependency` on the
`typescript` compiler and return to using just `tsc`. Also, pass the
`--build` argument, which speeds up incremental builds considerably.

Signed-off-by: 🧑🏻‍💻 Romain Marcadier <rmuller@amazon.com>