fix(import): allow `+` , `*` and `:` in enums (#1181)
During an import, `cdk8s-cli` checks for certain characters in `enum`s  and replaces values perceived as invalid with `STRIPPED_BY_CDK8S`. This fails for multiple CRDs in the wild which contain `+` and `:` in `enum` value. This PR fixes that. Generated code looks fine from my POV. 

Fixes #1072
Fixes #267