fix(import): incorrect and unnecessary properties recorder to `cdk8s.yaml` on `cdk8s import` (#1315)
When an import is specified with an alias, i.e `cdk8s import 1.24:=k8s@1.24`, the alias is being ignored when saving it to the configuration file. This creates duplicate import entries, one with the alias and one without. 

In addition, default configuration values that were only meant for in-process operation are written to the config file on import - this should not happen.

Also added a `--no-save` flag that skips writing the import to the file. This can be helpful if users have comments in `cdk8s.yaml`, which cannot be preserved upon rewrite. 

Fixes https://github.com/cdk8s-team/cdk8s/issues/1442