fix: helm chart url is considered as crd during helm synthesis (#1513)
Currently we are just ignoring `k8s` in `cdk8s.yaml` config file during synthesizing cdk8s app as a helm chart. This is because helm chart only needs `crds`(Custom Resource Definition) in it.

With the new url structure for importing `helm charts` i.e. prefixed with `helm:`, this also is being added to crds which throws an error when it tries to download it. For instance,

`'helm:https://lacework.github.io/helm-charts/lacework-agent@6.9.0'` URL would not lead to a CRD and hence the error.

This PR is ignoring the `helm:` prefix as well since that is not a CRD.

(cherry picked from commit 03a05cbebcce4aede922e3318c39bc27454bd97e)
Signed-off-by: Vinayak Kukreja <78971045+vinayak-kukreja@users.noreply.github.com>