Add YAML document separator to each manifest
When output to STDOUT is requested, each manifest in an app is rendered separately and dumped to standard output.
In case an app contains more than one chart, a document separator will be missing between adjacent charts, since there
is no YAML document separator either at the beginning or at the end of an individual output. The last resource of the
preceding chart and the first resource of the following chart will be merged into a random mess. By adding a YAML
document separator after each manifest unconditionally, this can be cleanly avoided.

Fixes: #943

Signed-off-by: Nikolai Prokoschenko <nikolai.prokoschenko@kurzdigital.com>