chore: prevent churning dependency on bleeding edge TypeScript (#3581)
Currently, the project has a transitive dependency on `typescript@next`:

```
cdk8s-cli@0.0.0
└─┬ jsii-srcmak@0.1.1236
  └─┬ jsii@5.5.33
    └─┬ downlevel-dts@0.11.0
      └── typescript@next
```

The `typescript@next` dependency is updated frequently by the automated
runtime dependency upgrades such as
https://github.com/cdk8s-team/cdk8s-cli/commit/243c142f223904c4616f90ddb6af4b44834ba46c
and
https://github.com/cdk8s-team/cdk8s-cli/commit/ecb03313e70bdb1aa7c4cc23db5b8f2a355a6c26.
Each of these causes a new `cdk8s-cli` release, which causes unnecessary
churn.

As a consumer of the `cdk8s-cli` project, this introduces a transitive
dependency for my project on `typescript@next`. With automated
dependency updates, my project ends up with the same churn where
`typescript@next` is constantly updated
(https://github.com/sandersn/downlevel-dts/issues/90).

The `downlevel-dts` package is no longer a dependency in more recent
`jsii` versions, which removes the `typescript@next` dependency. This
project currently uses a pinned `jsii-srcmak` version, due to the
configuration in
https://github.com/cdk8s-team/cdk8s-cli/commit/2577282a36b97d013afda38ecf93b05157dbaa62.

This change performs a manual upgrade for `jsii-srcmak` from `0.1.1236`
to `0.1.1319`, to remove the `typescript@next` dependency. I checked the
release notes for `jsii-srcmak` between these versions, and the only
changes are automated dependency updates. There don't appear to be any
new features or breaking changes reported in the release notes.

Signed-off-by: Daniel Frankcom <daniel@frankcom.ca>
Signed-off-by: Eli Polonsky <epolon@amazon.com>
Co-authored-by: Eli Polonsky <epolon@amazon.com>
Signed-off-by: Henning Schröder <henning.schroeder@gmail.com>